restructured authors
