+++ src/CoolBytes.WebAPI/Features/Authors/CQ/GetAuthorQuery.cs
@@ -1,6 +1,7 @@
-﻿using MediatR;
+﻿using CoolBytes.WebAPI.Features.Authors.ViewModels;
+using MediatR;
 
-namespace CoolBytes.WebAPI.Features.Authors
+namespace CoolBytes.WebAPI.Features.Authors.CQ
 {
     public class GetAuthorQuery : IRequest<AuthorViewModel>
     {
