+++ src/CoolBytes.WebAPI/Features/Authors/Handlers/AddAuthorCommandHandler.cs
@@ -1,13 +1,15 @@
-﻿using AutoMapper;
-using CoolBytes.Data;
-using MediatR;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using AutoMapper;
 using CoolBytes.Core.Abstractions;
 using CoolBytes.Core.Domain;
+using CoolBytes.Data;
+using CoolBytes.WebAPI.Features.Authors.CQ;
+using CoolBytes.WebAPI.Features.Authors.ViewModels;
+using MediatR;
 
-namespace CoolBytes.WebAPI.Features.Authors
+namespace CoolBytes.WebAPI.Features.Authors.Handlers
 {
     public class AddAuthorCommandHandler : IRequestHandler<AddAuthorCommand, AuthorViewModel>
     {
