+++ src/CoolBytes.WebAPI/Features/Authors/Profiles/SocialHandlesViewModelProfile.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using CoolBytes.Core.Domain;
+using CoolBytes.WebAPI.Features.Authors.ViewModels;
 
 namespace CoolBytes.WebAPI.Features.Authors.Profiles
 {
