+++ src/CoolBytes.WebAPI/Features/BlogPosts/ViewModels/BlogPostViewModel.cs
@@ -1,4 +1,4 @@
-﻿using CoolBytes.WebAPI.Features.Authors;
+﻿using CoolBytes.WebAPI.Features.Authors.ViewModels;
 using CoolBytes.WebAPI.Features.BlogPosts.DTO;
 using CoolBytes.WebAPI.Features.Images.ViewModels;
 using System;
