+++ src/CoolBytes.WebAPI/Features/Images/ViewModels/AzureBlobImageViewModelFactory.cs
@@ -1,6 +1,5 @@
 ﻿using CoolBytes.Core.Attributes;
 using CoolBytes.Core.Domain;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace CoolBytes.WebAPI.Features.Images.ViewModels
@@ -8,13 +7,6 @@ namespace CoolBytes.WebAPI.Features.Images.ViewModels
     [Inject(typeof(IImageViewModelFactory), ServiceLifetime.Scoped, "development", "production-azure")]
     public class AzureBlobImageViewModelFactory : IImageViewModelFactory
     {
-        private string _environment;
-
-        public AzureBlobImageViewModelFactory(IHostingEnvironment environment)
-        {
-            _environment = environment.EnvironmentName.ToLower();
-        }
-
         public ImageViewModel Create(Image image)
         {
             if (image == null)
